entity model: move claims utils to their own file
As recommended in https://github.com/inventaire/inventaire/pull/748#discussion_r1623183150
