+++ server/controllers/entities/lib/add_image_data.ts
@@ -1,10 +1,10 @@
 import { compact, pick } from 'lodash-es'
+import { getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { getWikimediaThumbnailData } from '#data/commons/thumb'
 // import getOpenLibraryCover from '#data/openlibrary/cover'
 // import getEnwikiImage from '#data/wikipedia/image'
 import { objectPromise } from '#lib/promises'
 // import { logError } from '#lib/utils/logs'
-import { getFirstClaimValue } from '#models/entity'
 import type { SerializedWdEntity, WikimediaCommonsFilename } from '#server/types/entity'
 import type { ImageData } from '#server/types/image'
 import { getCommonsFilenamesFromClaims } from './get_commons_filenames_from_claims.js'
