+++ server/controllers/entities/lib/create_wd_entity.ts
@@ -1,10 +1,10 @@
+import { getClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { getWikidataOAuthCredentials, validateWikidataOAuth } from '#controllers/entities/lib/wikidata_oauth'
 import { newError } from '#lib/error/error'
 import { mapKeysValues, objectEntries } from '#lib/utils/base'
 import { log } from '#lib/utils/logs'
 import { relocateQualifierProperties } from '#lib/wikidata/data_model_adapter'
 import wdEdit from '#lib/wikidata/edit'
-import { getClaimValue } from '#models/entity'
 import type { DatatypedInvClaimObject, EntityUri, EntityValue, ExpandedClaims, InvExpandedPropertyClaims, InvSnakValue, Labels, PropertyUri, Reference, WdEntityId, WdPropertyId } from '#server/types/entity'
 import type { User } from '#server/types/user'
 import { getEntityType } from './get_entity_type.js'
