+++ server/controllers/entities/lib/entities.ts
@@ -1,4 +1,5 @@
 import { uniqBy, cloneDeep, identity, pick, uniq } from 'lodash-es'
+import { getClaimValue, getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { authorRelationsProperties } from '#controllers/entities/lib/properties/properties'
 import dbFactory from '#db/couchdb/base'
 import { mapDoc } from '#lib/couch'
@@ -7,7 +8,7 @@ import { getUrlFromImageHash } from '#lib/images'
 import { toIsbn13h } from '#lib/isbn/isbn'
 import { emit } from '#lib/radio'
 import { assert_ } from '#lib/utils/assert_types'
-import { addEntityDocClaims, beforeEntityDocSave, getClaimValue, getFirstClaimValue, setEntityDocLabels } from '#models/entity'
+import { addEntityDocClaims, beforeEntityDocSave, setEntityDocLabels } from '#models/entity'
 import type { EntityImagePath, ImageHash } from '#server/types/image'
 import type { EntityUri, InvEntityDoc, EntityValue, PropertyUri, InvEntity, Isbn, InvClaimValue, SerializedEntity } from '#types/entity'
 import { getInvEntityCanonicalUri } from './get_inv_entity_canonical_uri.js'
