+++ server/controllers/entities/lib/format_edition_entity.ts
@@ -1,5 +1,5 @@
+import { getFirstClaimValue, simplifyInvClaims } from '#controllers/entities/lib/inv_claims_utils'
 import { normalizeIsbn } from '#lib/isbn/isbn'
-import { getFirstClaimValue, simplifyInvClaims } from '#models/entity'
 import type { InvEntity, IsbnEntityUri } from '#server/types/entity'
 import { formatEntityCommon } from './format_entity_common.js'
