+++ server/controllers/entities/lib/get_author_works.ts
@@ -1,5 +1,6 @@
 import { flatten, identity, map, uniqBy } from 'lodash-es'
 import { uniqByUri, getInvEntitiesByClaims, type ClaimPropertyValueTuple } from '#controllers/entities/lib/entities'
+import { getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { getEntitiesPopularities } from '#controllers/entities/lib/popularity'
 import { prefixifyWd } from '#controllers/entities/lib/prefix'
 import { authorRelationsProperties } from '#controllers/entities/lib/properties/properties'
@@ -7,7 +8,6 @@ import runWdQuery from '#data/wikidata/run_query'
 import { initCollectionsIndex } from '#lib/utils/base'
 import { LogErrorAndRethrow } from '#lib/utils/logs'
 import { getPluralType, getPluralTypeByTypeUri } from '#lib/wikidata/aliases'
-import { getFirstClaimValue } from '#models/entity'
 import type { ViewRow } from '#server/types/couchdb'
 import type { EntityUri, EntityValue, InvEntity } from '#server/types/entity'
 import { getSimpleDayDate, sortByScore } from './queries_utils.js'
