+++ server/controllers/entities/lib/get_inv_entities.ts
@@ -2,8 +2,8 @@ import { difference } from 'lodash-es'
 import { getEntitiesByIds } from '#controllers/entities/lib/entities'
 import type { EntitiesGetterParams } from '#controllers/entities/lib/get_entities_by_uris'
 import { getEntityByUri } from '#controllers/entities/lib/get_entity_by_uri'
+import { simplifyInvClaims } from '#controllers/entities/lib/inv_claims_utils'
 import { prefixifyInv, unprefixify } from '#controllers/entities/lib/prefix'
-import { simplifyInvClaims } from '#models/entity'
 import type { EntityRedirection, InvEntityDoc, InvEntityId, InvEntityUri, SerializedEntity, SerializedInvEntity, SerializedRemovedPlaceholder } from '#types/entity'
 import { formatEntityCommon } from './format_entity_common.js'
 import { getEntityType } from './get_entity_type.js'
