+++ server/controllers/entities/lib/get_inv_uri_from_doc.ts
@@ -1,5 +1,5 @@
+import { getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { normalizeIsbn } from '#lib/isbn/isbn'
-import { getFirstClaimValue } from '#models/entity'
 import type { InvEntity, IsbnEntityUri, RemovedPlaceholderEntity } from '#server/types/entity'
 import { prefixifyInv } from './prefix.js'
