+++ server/controllers/entities/lib/get_publisher_publications.ts
@@ -1,7 +1,7 @@
 import { getInvEntitiesByClaim } from '#controllers/entities/lib/entities'
+import { getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { prefixifyWd } from '#controllers/entities/lib/prefix'
 import runWdQuery from '#data/wikidata/run_query'
-import { getFirstClaimValue } from '#models/entity'
 import type { InvEntity, WdEntityId, WdEntityUri } from '#server/types/entity'
 import { getInvEntityCanonicalUri } from './get_inv_entity_canonical_uri.js'
 import type { Split } from 'type-fest'
