+++ server/controllers/entities/lib/get_serie_parts.ts
@@ -1,9 +1,9 @@
 import { flatten } from 'lodash-es'
 import { getInvEntitiesByClaim, uniqByUri } from '#controllers/entities/lib/entities'
+import { getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { prefixifyWd } from '#controllers/entities/lib/prefix'
 import runWdQuery from '#data/wikidata/run_query'
 import { LogErrorAndRethrow } from '#lib/utils/logs'
-import { getFirstClaimValue } from '#models/entity'
 import type { EntityUri, InvEntity, SerializedEntity, WdEntityId } from '#server/types/entity'
 import { getSimpleDayDate, sortByOrdinalOrDate } from './queries_utils.js'
 import { getCachedRelations } from './temporarily_cache_relations.js'
