+++ server/controllers/entities/lib/get_wikidata_enriched_entities.ts
@@ -7,6 +7,7 @@
 import { partition, map, compact, omit } from 'lodash-es'
 import { simplifyAliases, simplifyDescriptions, simplifyLabels, simplifyPropertyClaims, simplifySitelinks, type Claims, type Item as WdEntity } from 'wikibase-sdk'
 import type { EntitiesGetterParams } from '#controllers/entities/lib/get_entities_by_uris'
+import { getClaimObjectFromClaim } from '#controllers/entities/lib/inv_claims_utils'
 import { prefixifyWd, unprefixify } from '#controllers/entities/lib/prefix'
 import { getWdEntity } from '#data/wikidata/get_entity'
 import { hardCodedUsers } from '#db/couchdb/hard_coded_documents'
@@ -16,7 +17,6 @@ import { emit } from '#lib/radio'
 import { objectEntries } from '#lib/utils/base'
 import { formatClaims } from '#lib/wikidata/format_claims'
 import getOriginalLang from '#lib/wikidata/get_original_lang'
-import { getClaimObjectFromClaim } from '#models/entity'
 import type { ExtendedEntityType, ExpandedSerializedWdEntity, SerializedWdEntity, WdEntityId, WdEntityUri } from '#types/entity'
 import { addImageData } from './add_image_data.js'
 import { getEntityType } from './get_entity_type.js'
