+++ server/controllers/entities/lib/resolver/create_entity_from_seed.ts
@@ -1,10 +1,10 @@
-import { chain, compact, map, without } from 'lodash-es'
+import { chain, compact, map } from 'lodash-es'
+import { findClaimByValue, getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { convertAndCleanupImageUrl } from '#controllers/images/lib/convert_and_cleanup_image_url'
 import { getImageByIsbn } from '#data/dataseed/dataseed'
 import { isNonEmptyString } from '#lib/boolean_validations'
 import { toIsbn13h } from '#lib/isbn/isbn'
 import { logError, warn } from '#lib/utils/logs'
-import { findClaimByValue, getFirstClaimValue } from '#models/entity'
 import type { Claims, EntityType, InvSimplifiedPropertyClaims, PropertyUri } from '#types/entity'
 import type { BatchId } from '#types/patch'
 import type { EditionSeed, EntitySeed } from '#types/resolver'
