+++ server/controllers/entities/lib/resolver/resolve_edition_from_works.ts
@@ -1,8 +1,8 @@
 import { getInvEntitiesByClaim } from '#controllers/entities/lib/entities'
 import { getInvEntityCanonicalUri } from '#controllers/entities/lib/get_inv_entity_canonical_uri'
+import { findClaimByValue, getClaimValue, getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { normalizeTitle } from '#controllers/entities/lib/resolver/helpers'
 import { objectEntries } from '#lib/utils/base'
-import { findClaimByValue, getClaimValue, getFirstClaimValue } from '#models/entity'
 import type { EditionSeed, WorkSeed } from '#server/types/resolver'
 import type { InvEntity } from '#types/entity'
