+++ server/controllers/entities/lib/resolver/resolve_external_ids.ts
@@ -1,9 +1,9 @@
 import { flatten, uniq } from 'lodash-es'
 import { getInvEntitiesByClaim } from '#controllers/entities/lib/entities'
+import { getClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { prefixifyWd } from '#controllers/entities/lib/prefix'
 import runWdQuery from '#data/wikidata/run_query'
 import { forceArray, objectEntries } from '#lib/utils/base'
-import { getClaimValue } from '#models/entity'
 import type { Claims } from '#server/types/entity'
 import { getInvEntityCanonicalUri } from '../get_inv_entity_canonical_uri.js'
 import { propertiesValuesConstraints as properties } from '../properties/properties_values_constraints.js'
