+++ server/controllers/entities/lib/resolver/sanitize_entry.ts
@@ -1,10 +1,10 @@
 import { isPlainObject } from 'lodash-es'
+import { getClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { isArray } from '#lib/boolean_validations'
 import { newError } from '#lib/error/error'
 import { guessLangFromIsbn, isValidIsbn, normalizeIsbn } from '#lib/isbn/isbn'
 import { forceArray } from '#lib/utils/base'
 import { requireJson } from '#lib/utils/json'
-import { getClaimValue, getFirstClaimValue } from '#models/entity'
 import type { EntityType, Isbn } from '#server/types/entity'
 import type { EditionLooseSeed, EntityLooseSeed, ResolverEntry, SanitizedResolverEntry } from '#server/types/resolver'
 import { sanitizeSeed } from './sanitize_seed.js'
