+++ server/controllers/entities/lib/resolver/update_resolved_entry.ts
@@ -1,9 +1,10 @@
 import { cloneDeep, isEqual, omit, pick } from 'lodash-es'
+import { getClaimValue, getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { normalizeTitle } from '#controllers/entities/lib/resolver/helpers'
 import type { ResolverBatchParams } from '#controllers/entities/lib/resolver/resolve_update_and_create'
 import { convertAndCleanupImageUrl } from '#controllers/images/lib/convert_and_cleanup_image_url'
 import { objectKeys } from '#lib/utils/types'
-import { addEntityDocClaims, getClaimValue, getFirstClaimValue } from '#models/entity'
+import { addEntityDocClaims } from '#models/entity'
 import type { Url } from '#server/types/common'
 import type { ClaimByDatatype, Claims, EntityId, EntityUriPrefix, InvEntity, InvEntityDoc, InvEntityId, Isbn } from '#server/types/entity'
 import type { BatchId } from '#server/types/patch'
