+++ server/controllers/entities/lib/temporarily_cache_relations.ts
@@ -1,9 +1,9 @@
 import { compact, map, uniq } from 'lodash-es'
 import { getAggregatedPropertiesValues, getEntityById } from '#controllers/entities/lib/entities'
 import { getEntitiesList } from '#controllers/entities/lib/get_entities_list'
+import { getClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { propertiesValuesConstraints, getPropertyDatatype } from '#controllers/entities/lib/properties/properties_values_constraints'
 import { objectKeys } from '#lib/utils/types'
-import { getClaimValue } from '#models/entity'
 import type { EntityUri, InvEntityUri, PropertyUri, SerializedEntity } from '#types/entity'
 import entitiesRelationsTemporaryCache from './entities_relations_temporary_cache.js'
 import { unprefixify } from './prefix.js'
