+++ server/controllers/entities/lib/turn_into_redirection.ts
@@ -1,9 +1,10 @@
 import { chain, compact } from 'lodash-es'
 import { getInvClaimsByClaimValue, getEntityById, putInvEntityUpdate } from '#controllers/entities/lib/entities'
+import { getClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { removePlaceholder } from '#controllers/entities/lib/placeholders'
 import { assert_ } from '#lib/utils/assert_types'
 import { log } from '#lib/utils/logs'
-import { convertEntityDocIntoARedirection, getClaimValue, preventRedirectionEdit } from '#models/entity'
+import { convertEntityDocIntoARedirection, preventRedirectionEdit } from '#models/entity'
 import type { Claims, EntityUri, InvEntity, InvEntityId, InvEntityUri, PropertyUri } from '#types/entity'
 import type { PatchContext } from '#types/patch'
 import type { UserId } from '#types/user'
