+++ server/controllers/entities/lib/validate_and_format_claim.ts
@@ -1,9 +1,9 @@
+import { isClaimObject } from '#controllers/entities/lib/inv_claims_utils'
 import { externalIdsProperties } from '#controllers/entities/lib/properties/properties_values_constraints'
 import { validateSnakValueSync } from '#controllers/entities/lib/validate_claim_sync'
 import { isNonEmptyPlainObject } from '#lib/boolean_validations'
 import { newError } from '#lib/error/error'
 import { arrayIncludes, objectEntries } from '#lib/utils/base'
-import { isClaimObject } from '#models/entity'
 import type { InvClaimObject } from '#server/types/entity'
 import { validateProperty } from './properties/validations.js'
 import { validateAndFormatClaimValue, type ValidateAndFormatClaimValueParams } from './validate_and_format_claim_value.js'
