+++ server/controllers/entities/merge.ts
@@ -1,8 +1,8 @@
+import { getFirstClaimValue } from '#controllers/entities/lib/inv_claims_utils'
 import { isIsbnEntityUri } from '#lib/boolean_validations'
 import { newError } from '#lib/error/error'
 import { emit } from '#lib/radio'
 import { log } from '#lib/utils/logs'
-import { getFirstClaimValue } from '#models/entity'
 import type { EntityUri, SerializedEntity } from '#types/entity'
 import { getEntitiesByUris } from './lib/get_entities_by_uris.js'
 import mergeEntities from './lib/merge_entities.js'
