+++ server/db/elasticsearch/formatters/entity.ts
@@ -5,6 +5,7 @@ import { imageProperties } from '#controllers/entities/lib/get_commons_filenames
 import { getEntitiesList } from '#controllers/entities/lib/get_entities_list'
 import { getEntityImagesFromClaims } from '#controllers/entities/lib/get_entity_images_from_claims'
 import { getEntityType } from '#controllers/entities/lib/get_entity_type'
+import { simplifyInvClaims } from '#controllers/entities/lib/inv_claims_utils'
 import { languagesCodesProperties } from '#controllers/entities/lib/languages'
 import { getEntityPopularity } from '#controllers/entities/lib/popularity'
 import { authorRelationsProperties } from '#controllers/entities/lib/properties/properties'
@@ -15,7 +16,6 @@ import { objectEntries } from '#lib/utils/base'
 import { warn } from '#lib/utils/logs'
 import { getSingularTypes } from '#lib/wikidata/aliases'
 import { formatClaims } from '#lib/wikidata/format_claims'
-import { simplifyInvClaims } from '#models/entity'
 import type { Claims, EntityUri, PropertyUri, SerializedEntity, WdRawClaims } from '#types/entity'
 import { activeI18nLangs } from '../helpers.js'
 import { getEntityId } from './entity_helpers.js'
